// src/core/types.ts
export interface ParserFeature {
    tags: string[];
    keyword: string;
    name: string;
    description: string;
    location: number;
}

interface ParserPolicy {
    type: 'undeclared-field' | 'required-field';
    field: string;
}

export interface ParserScenario {
    uri: string;
    feature: ParserFeature;
    location: number;
    keyword: string;
    name: string;
    description: string;
    steps: ParserStep[];
    tags: string[];
    background?: ParserBackground;
    rule?: ParserRule;
    custom?: {
        testcases?: string[];
        testruns?: string[];
        identity?: string;
        fields?: Record<string, string>;
        policy?: ParserPolicy[];
        groupScenarios?: ParserScenario[];
    };
}

export interface ParserBackground {
    keyword: string;
    name: string;
    steps: ParserStep[];
}

export interface ParserStep {
    keyword: string;
    keywordType?: string;
    text: string;
}

export interface ParserRule {
    keyword: string;
    name: string;
    description: string;
}
// ...
export interface PolicyRule {
    id: string;
    type?: 'error' | 'warning';
    title: string;
    detail: string;
    uri: string;
    scenario?: string;
    line?: number;
}